change quantity byte to short and add half part addProduct in UI

+++ OnlineShop/OnlineShop.Domain/AbstractRepository/ICartWriteRepository.cs
@@ -9,6 +9,6 @@ namespace OnlineShop.Domain.AbstractRepository
 
         Task RemoveCartAsync(long id);
 
-        Task UpdateCartQuantityAsync(long id, long productId, byte quantity);
+        Task UpdateCartQuantityAsync(long id, long productId, short quantity);
     }
 }
