+++ OnlineShop/OnlineShop.Domain/Carts/Commands/AddCartCommand.cs
@@ -10,10 +10,10 @@ namespace OnlineShop.Domain.Carts.Commands
     public sealed class AddCartCommand : Command
     {
         public long ProductId { get; set; }
-        public byte Quantity { get; set; }
+        public short Quantity { get; set; }
         public string UserId { get; set; }
 
-        public AddCartCommand(CommandMeta commandMeta, long productId, byte quantity, string userId) : base(commandMeta)
+        public AddCartCommand(CommandMeta commandMeta, long productId, short quantity, string userId) : base(commandMeta)
         {
             ProductId = productId;
             Quantity = quantity;
