+++ OnlineShop/OnlineShop.Domain/Carts/Entity/Cart.Model.cs
@@ -5,6 +5,6 @@
         public override long Id { get; protected set; }
         public long ProductId { get; private set; }
         public string UserId { get; private set; }
-        public byte Quantity { get; private set; }
+        public short Quantity { get; private set; }
     }
 }
