+++ OnlineShop/OnlineShop.Domain/Products/Commands/CreateProductCommand.cs
@@ -19,7 +19,7 @@ namespace OnlineShop.Domain.Products.Commands
         public Gender? Gender { get; }
         public bool? ForBaby { get; }
         public string? Size { get; }
-        public byte Quantity { get; }
+        public short Quantity { get; }
         public float? Discount { get; }
         public DateTime? CreateTime { get; }
         public List<ProductImage> Images { get; }
@@ -36,7 +36,7 @@ namespace OnlineShop.Domain.Products.Commands
             Gender? gender,
             bool? forBaby,
             string? size,
-            byte quantity,
+            short quantity,
             float? discount,
             DateTime? createTime,
             List<ProductImage> images,
